Change type of column-names option
Change type of column-names map option to map(xs:string, xs:integer)
from map(xs:integer, xs:string). It turns out I had used this format for
all the examples, and the more I thought about it, the more it seemed
unhelpful to have a map like `map { 1: "a", 2: "b" }` produce a
`csv-columns-record` whose `names` entry was `map { "a": 1, "b": 2 }`.

Not least, that prevented the `names` entry from a `csv-columns-record`
being used as the `column-names` option in another invocation, which
might be desirable if a user has one CSV with headers and subsequent
CSVs of the same schema but without the header line (transaction logs
split for size, perhaps).