Correction in `SimpleMapTarget` to avoid ambiguity
As @michaelhkay pointed out, there is an ambiguity,
since `RelativePathTarget` can be part of `ArrowExpr`
both directly and indirectly (via `SimpleMapTarget`).

By this change, `SimpleMapTarget` cannot contain `RelativePathTarget`.

It is probably not perfect that the terminal `!`
(for the simple map operator) is directly used in two
now independent productions: `SimpleMapExpr` and `SimpleMapTarget`.